QPID-7763: Correct return type of getAllocatedDirectMemorySize from int to long

+++ common/src/main/java/org/apache/qpid/bytebuffer/QpidByteBuffer.java
@@ -824,7 +824,7 @@ public class QpidByteBuffer
         return _pooledBufferSize;
     }
 
-    public static int getAllocatedDirectMemorySize()
+    public static long getAllocatedDirectMemorySize()
     {
         return _pooledBufferSize * getNumberOfActivePooledBuffers();
     }
